navigation to app detail working (slug workaround)
I'm not happy with the slug workaround. I guess the global unique id
should actually be in the url. But that does not seem to work. I think
the example in relay-starter-kit actually handles the lookup for a
specific item wrong. It should take the globally unique id and work. Not
the type-specific id.